requirements.txt
absl-py==1.4.0
anx-interface==0.5.3
attrs==23.1.0
cffi==1.15.1
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
evdev==1.6.1
flatbuffers==23.5.26
fonttools==4.42.0
importlib-resources==6.0.1
kiwisolver==1.4.4
matplotlib==3.7.2
mediapipe==0.10.3
numpy==1.24.4
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
packaging==23.1
Pillow==9.5.0
progress==1.6
protobuf==3.12.4
psutil==5.9.5
pybind11==2.11.1
pycparser==2.21
pynput==1.7.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-xlib==0.33
pyzmq==22.3.0
six==1.16.0
sounddevice==0.4.6
zipp==3.16.2
